refactor: rename Hero component to HeroComponent to disambiguate Heros
The generated graphql.tsx got type Hero from CMS and that clashed with
the function Hero in components.
