+++ packages/components/src/components/hero/heroComponent.module.scss
